feat: 修复扫描禁用制品失败 #3947
1. 修复手动禁用也会被扫描方案解禁问题
2. 设置扫描失败时保持制品禁用状态
3. 修复扫描禁用制品失败

+++ src/backend/analyst/biz-analyst/src/main/kotlin/com/tencent/bkrepo/analyst/event/listener/SubtaskStatusChangedEventListener.kt
@@ -37,6 +37,7 @@ import com.tencent.bkrepo.analyst.service.ScanQualityService
 import com.tencent.bkrepo.common.analysis.pojo.scanner.SubScanTaskStatus
 import com.tencent.bkrepo.common.analysis.pojo.scanner.standard.StandardScanExecutorResult
 import com.tencent.bkrepo.common.api.util.toJsonString
+import com.tencent.bkrepo.common.artifact.constant.FORBID_STATUS
 import com.tencent.bkrepo.common.artifact.constant.FORBID_TYPE
 import com.tencent.bkrepo.common.artifact.constant.SCAN_STATUS
 import com.tencent.bkrepo.common.artifact.pojo.RepositoryType
@@ -168,14 +169,8 @@ class SubtaskStatusChangedEventListener(
         subtask.qualityRedLine?.let {
             metadata.add(MetadataModel(key = SubScanTaskDefinition::qualityRedLine.name, value = it, system = true))
         }
-        val currentForbidType = with(subtask) {
-            if (repoType == RepositoryType.GENERIC.name) {
-                metadataService.listMetadata(projectId, repoName, fullPath)[FORBID_TYPE]
-            } else if (!packageKey.isNullOrEmpty() && !version.isNullOrEmpty()) {
-                packageMetadataService.listMetadata(projectId, repoName, packageKey, version)[FORBID_TYPE]
-            }
-        }
-        if (currentForbidType == ForbidType.MANUAL.name) {
+
+        if (manualForbidden(subtask)) {
             // 手动禁用的情况不处理
             return
         }
@@ -187,7 +182,7 @@ class SubtaskStatusChangedEventListener(
             val user = result.plan?.lastModifiedBy ?: SYSTEM_USER
             metadata.addAll(MetadataUtils.generateForbidMetadata(true, type.reason, type, user))
             return
-        } else {
+        } else if (result.shouldUnforbid) {
             // 全部方案均通过时移除禁用元数据
             deleteMetadata(subtask, MetadataUtils.FORBID_KEYS)
         }
@@ -201,6 +196,23 @@ class SubtaskStatusChangedEventListener(
             null
         }
         return SubtaskRecord(subtask = event.subtask, metadata = event.taskMetadata, metrics = metrics)
+    }
+
+    private fun manualForbidden(subtask: TPlanArtifactLatestSubScanTask): Boolean {
+        with(subtask) {
+            val metadata = if (repoType == RepositoryType.GENERIC.name) {
+                metadataService.listMetadata(projectId, repoName, fullPath)
+            } else if (!packageKey.isNullOrEmpty() && !version.isNullOrEmpty()) {
+                packageMetadataService.listMetadata(projectId, repoName, packageKey, version)
+            } else {
+                emptyMap()
+            }
+
+            val forbidType = metadata[FORBID_TYPE]
+            val forbidStatus = metadata[FORBID_STATUS]
+
+            return forbidStatus == true && forbidType == ForbidType.MANUAL.name
+        }
     }
 
     private data class SubtaskRecord(
